+++ CHANGELOG.md
@@ -9,8 +9,13 @@
 
 - Properties on `CameraModel` (e.g. `cam.M`) are deprecated and will
   be removed in a future release.
+
 - Intrinsic parameter matrix is normalized upon loading in
   `CameraModel._from_parts`, which is called by most constructors.
+
+- If an input calibration has a non-normalized P matrix and a
+  rectification matrix, a warning is given as this case is not well
+  tested and the behavior should be considered undefined.
 
 ## Bugfixes